Ported Header login div to System module
Also included some formatting changes to User model to fix some syntax highlighting issues in Sublime. Added the login controller to the header.js System file. Most of the functionality is there, however I'm still manually reloading on login. The eventual goal is to have no reloads, but this is a good start